requirements.txt
streamlit
pandas
numpy
huggingface_hub
requests
plotly
matplotlib
pydantic
python-dotenv
tqdm
rich
SQLAlchemy
transformers


